requirements.txt
annotated-types==0.7.0
audioread==3.1.0
certifi==2022.12.7
cffi==2.0.0
charset-normalizer==2.1.1
cmake==3.25.0
decorator==5.2.1
deepspeed==0.9.0
einops==0.8.1
filelock==3.19.1
fsspec==2025.10.0
hf-xet==1.2.0
hjson==3.1.0
huggingface-hub==0.36.0
idna==3.4
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
lit==15.0.7
llvmlite==0.45.1
MarkupSafe==2.1.5
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
networkx==3.3
ninja==1.13.0
numba==0.62.1
numpy==1.26.4
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
progressbar==2.5
psutil==7.1.2
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
regex==2025.10.23
requests==2.28.1
rotary-embedding-torch==0.8.9
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tortoise-tts==3.0.0
tqdm==4.67.1
transformers==4.31.0
triton==2.0.0
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==1.26.13
